refactor(language-service): ♻️ extract hover docs logic into separate file
